+++ hub/src/main/java/com/sparta/hub/domain/route/service/HubRouteService.java
@@ -1,14 +1,11 @@
 package com.sparta.hub.domain.route.service;
 
-import com.sparta.hub.application.exception.ErrorCode;
-import com.sparta.hub.application.exception.ServiceException;
 import com.sparta.hub.domain.hub.model.Hub;
 import com.sparta.hub.domain.route.dto.HubRoutePatchDto;
 import com.sparta.hub.domain.route.model.HubRoute;
 import com.sparta.hub.domain.route.model.RouteInfo;
 import com.sparta.hub.domain.route.service.cache.HubRouteCacheService;
-import com.sparta.hub.domain.route.service.utils.RouteSegmentCalculator;
-import com.sparta.hub.domain.route.service.utils.RouteEstimator;
+import com.sparta.hub.domain.route.service.utils.RouteCalculationService;
 import com.sparta.hub.infrastructure.persistence.HubRepository;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.HashSet;
@@ -32,53 +29,12 @@ public class HubRouteService {
 
     private final HubRepository hubRepository;
     private final HubRouteCacheService hubRouteCacheService;
-
-    private final RouteEstimator routeEstimator;
-    private final RouteSegmentCalculator routeSegmentCalculator;
+    private final RouteCalculationService routeCalculationService;
 
     public HubRoute createHubRoute(UUID originHubId, UUID destinationHubId) {
         log.info("Creating hub route from {} to {}", originHubId, destinationHubId);
-
-        List<UUID> calculatedRouteSegments;
-        try {
-            calculatedRouteSegments = routeSegmentCalculator.calculateRouteSegments(originHubId, destinationHubId);
-            log.info("Calculated route segments: {}", calculatedRouteSegments);
-        } catch (Exception e) {
-            log.error("Failed to calculate route segments: {}", e.getMessage());
-            throw new ServiceException(ErrorCode.INTERNAL_SERVER_ERROR);
-        }
-
-        Map<UUID, Hub> hubMap;
-        try {
-            hubMap = fetchRequiredHubs(calculatedRouteSegments);
-        } catch (ServiceException e) {
-            log.error("Failed to fetch required hubs: {}", e.getMessage());
-            throw e;
-        }
-
         HubRoute hubRoute = HubRoute.create(originHubId, destinationHubId);
-        hubRoute.updateRouteSegments(calculatedRouteSegments);
-
-        RouteInfo routeInfo;
-        try {
-            routeInfo = calculateRouteInfo(calculatedRouteSegments, hubMap);
-        } catch (ServiceException e) {
-            log.error("Failed to calculate route info: {}", e.getMessage());
-            throw e;
-        }
-
-        String routeDisplayName = generateRouteDisplayName(originHubId, destinationHubId, hubMap);
-
-        hubRoute.updateHubRouteInfo(routeInfo.getTotalEstimatedTime(),
-                routeInfo.getTotalEstimatedDistance(),
-                routeDisplayName);
-
-        try {
-            return hubRouteCacheService.saveHubRoute(hubRoute);
-        } catch (Exception e) {
-            log.error("Failed to save hub route: {}", e.getMessage());
-            throw new ServiceException(ErrorCode.INTERNAL_SERVER_ERROR);
-        }
+        return updateRouteInformation(hubRoute, originHubId, destinationHubId);
     }
 
     @Transactional
@@ -94,40 +50,7 @@ public class HubRouteService {
 
         if (isRouteChanged) {
             log.info("Route changed. Recalculating route segments.");
-            List<UUID> newRouteSegments;
-            try {
-                newRouteSegments = routeSegmentCalculator.calculateRouteSegments(newOriginHubId, newDestinationHubId);
-                log.info("Calculated new route segments: {}", newRouteSegments);
-            } catch (Exception e) {
-                log.error("Failed to calculate new route segments: {}", e.getMessage());
-                throw new ServiceException(ErrorCode.INTERNAL_SERVER_ERROR);
-            }
-
-            Map<UUID, Hub> hubMap;
-            try {
-                hubMap = fetchRequiredHubs(newRouteSegments);
-            } catch (ServiceException e) {
-                log.error("Failed to fetch required hubs: {}", e.getMessage());
-                throw e;
-            }
-
-            RouteInfo routeInfo;
-            try {
-                routeInfo = calculateRouteInfo(newRouteSegments, hubMap);
-            } catch (ServiceException e) {
-                log.error("Failed to calculate route info: {}", e.getMessage());
-                throw e;
-            }
-
-            String newRouteDisplayName = generateRouteDisplayName(newOriginHubId, newDestinationHubId, hubMap);
-
-            existingRoute.updateOriginAndDestination(newOriginHubId, newDestinationHubId);
-            existingRoute.updateRouteSegments(newRouteSegments);
-            existingRoute.updateHubRouteInfo(
-                    routeInfo.getTotalEstimatedTime(),
-                    routeInfo.getTotalEstimatedDistance(),
-                    newRouteDisplayName
-            );
+            return updateRouteInformation(existingRoute, newOriginHubId, newDestinationHubId);
         } else {
             log.info("Route not changed. Updating other information.");
             existingRoute.updateHubRouteInfo(
@@ -135,54 +58,60 @@ public class HubRouteService {
                     hubRoutePatchDto.getEstimatedDistance(),
                     hubRoutePatchDto.getRouteDisplayName()
             );
-        }
-
-        try {
             return hubRouteCacheService.saveHubRoute(existingRoute);
-        } catch (Exception e) {
-            log.error("Failed to save updated hub route: {}", e.getMessage());
-            throw new ServiceException(ErrorCode.INTERNAL_SERVER_ERROR);
         }
     }
 
+    private HubRoute updateRouteInformation(HubRoute hubRoute, UUID originHubId, UUID destinationHubId) {
+        List<UUID> calculatedRouteSegments = calculateRouteSegments(originHubId, destinationHubId);
+        Map<UUID, Hub> hubMap = fetchRequiredHubs(calculatedRouteSegments);
+        RouteInfo routeInfo = calculateRouteInfo(calculatedRouteSegments, hubMap);
+        String routeDisplayName = generateRouteDisplayName(originHubId, destinationHubId, hubMap);
+
+        hubRoute.updateOriginAndDestination(originHubId, destinationHubId);
+        hubRoute.updateRouteSegments(calculatedRouteSegments);
+        hubRoute.updateHubRouteInfo(
+                routeInfo.getTotalEstimatedTime(),
+                routeInfo.getTotalEstimatedDistance(),
+                routeDisplayName
+        );
+
+        return hubRouteCacheService.saveHubRoute(hubRoute);
+    }
+
+    private List<UUID> calculateRouteSegments(UUID originHubId, UUID destinationHubId) {
+        return routeCalculationService.calculateRouteSegments(originHubId, destinationHubId);
+    }
+
     private Map<UUID, Hub> fetchRequiredHubs(List<UUID> hubIds) {
-        List<Hub> hubs = hubRepository.findAllById(hubIds);
-        Map<UUID, Hub> hubMap = hubs.stream().collect(Collectors.toMap(Hub::getHubId, Function.identity()));
 
-        Set<UUID> notFoundHubIds = new HashSet<>(hubIds);
-        notFoundHubIds.removeAll(hubMap.keySet());
+        validateHubs(hubIds.toArray(UUID[]::new));
 
-        if (!notFoundHubIds.isEmpty()) {
-            log.error("다음 ID를 가진 허브를 찾을 수 없습니다: {}", notFoundHubIds);
-            throw new ServiceException(ErrorCode.HUB_NOT_FOUND);
+        List<Hub> hubs = hubRepository.findByIdsAndIsDeletedFalse(hubIds);
+        Map<UUID, Hub> hubMap = hubs.stream()
+                .collect(Collectors.toMap(Hub::getHubId, Function.identity()));
+
+        // 조회된 허브 수와 요청된 ID 수가 일치하는지 확인
+        if (hubMap.size() != hubIds.size()) {
+            log.warn("Number of fetched hubs ({}) does not match the number of requested IDs ({})",
+                    hubMap.size(), hubIds.size());
+            Set<UUID> missingHubIds = new HashSet<>(hubIds);
+            missingHubIds.removeAll(hubMap.keySet());
+            throw new EntityNotFoundException("Some hubs were not found or are deleted: " + missingHubIds);
         }
 
         return hubMap;
     }
 
     private RouteInfo calculateRouteInfo(List<UUID> routeSegments, Map<UUID, Hub> hubMap) {
-        int totalEstimatedTime = 0;
-        double totalEstimatedDistance = 0;
-
-        for (int i = 0; i < routeSegments.size() - 1; i++) {
-            UUID currentHubId = routeSegments.get(i);
-            UUID nextHubId = routeSegments.get(i + 1);
-
-            Hub currentHub = hubMap.get(currentHubId);
-            Hub nextHub = hubMap.get(nextHubId);
-
-            if (currentHub == null || nextHub == null) {
-                log.error("Hub not found. Current hub ID: {}, Next hub ID: {}", currentHubId, nextHubId);
-                throw new ServiceException(ErrorCode.HUB_NOT_FOUND);
-            }
-
-            totalEstimatedTime += routeEstimator.calculateEstimatedTime(currentHub, nextHub);
-            totalEstimatedDistance += routeEstimator.calculateDistance(currentHub, nextHub);
-        }
-
-        return new RouteInfo(totalEstimatedTime, totalEstimatedDistance);
+        return routeCalculationService.calculateRouteInfo(routeSegments, hubMap);
     }
 
+    private String generateRouteDisplayName(UUID originHubId, UUID destinationHubId, Map<UUID, Hub> hubMap) {
+        Hub originHub = hubMap.get(originHubId);
+        Hub destinationHub = hubMap.get(destinationHubId);
+        return String.format("%s → %s", originHub.getName(), destinationHub.getName());
+    }
 
     public void deleteHubRoute(UUID hubRouteId) {
         hubRouteCacheService.deleteHubRoute(hubRouteId);
@@ -190,6 +119,11 @@ public class HubRouteService {
 
     public HubRoute getHubRoute(UUID hubRouteId) {
         return hubRouteCacheService.getHubRoute(hubRouteId);
+    }
+
+    public HubRoute getHubRouteByOriginAndDestination(UUID originHubId, UUID destinationHubId) {
+        validateHubs(originHubId, destinationHubId);
+        return hubRouteCacheService.getHubRoutesByOriginAndDestination(originHubId, destinationHubId);
     }
 
     public List<HubRoute> getHubRoutesByOrigin(UUID originHubId) {
@@ -205,18 +139,6 @@ public class HubRouteService {
     public Page<HubRoute> getAllHubRoutes(Pageable pageable) {
         return hubRouteCacheService.getAllHubRoutes(pageable);
     }
-
-    private String generateRouteDisplayName(UUID originHubId, UUID destinationHubId, Map<UUID, Hub> hubMap) {
-        Hub originHub = hubMap.get(originHubId);
-        Hub destinationHub = hubMap.get(destinationHubId);
-
-        if (originHub == null || destinationHub == null) {
-            throw new ServiceException(ErrorCode.HUB_NOT_FOUND);
-        }
-
-        return originHub.getName() + " -> " + destinationHub.getName();
-    }
-
 
     private void validateHub(UUID hubId) {
         if (!hubRepository.existsById(hubId)) {
